+++ panel/src/components/View/Buttons/index.js
@@ -1,4 +1,4 @@
-import ContentTranslations from "./ContentTranslationsButton.vue";
+import LanguagesDropdown from "./LanguagesDropdown.vue";
 import Settings from "./SettingsButton.vue";
 import Status from "./StatusButton.vue";
 import Theme from "./ThemeButton.vue";
@@ -8,15 +8,12 @@ import Buttons from "./Buttons.vue";
 
 export default {
 	install(app) {
-		app.component("k-content-translations-view-button", ContentTranslations);
+		app.component("k-languages-dropdown", LanguagesDropdown);
 		app.component("k-settings-view-button", Settings);
 		app.component("k-status-view-button", Status);
 		app.component("k-theme-view-button", Theme);
 
 		app.component("k-view-button", Button);
 		app.component("k-view-buttons", Buttons);
-
-		// @deprecated
-		app.component("k-languages-dropdown", ContentTranslations);
 	}
 };
